fixed bug with guest props timestamp
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@12512 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/ConsoleImpl.cpp
@@ -3611,7 +3611,7 @@ HRESULT Console::getGuestProperty (INPTR BSTR aName, BSTR *aValue,
         {
             size_t iFlags = strlen(pszBuffer) + 1;
             Utf8Str(pszBuffer).cloneTo (aValue);
-            *aTimestamp = parm[2].u.uint32;
+            *aTimestamp = parm[2].u.uint64;
             Utf8Str(pszBuffer + iFlags).cloneTo (aFlags);
         }
         else
